updated LibFlatArray

+++ lib/libflatarray/include/libflatarray/detail/save_functor.hpp
@@ -39,10 +39,10 @@ public:
 
         for (ITERATOR i = start; i != end; ++i) {
             accessor.index() = soa_accessor<CELL, DIM_X, DIM_Y, DIM_Z, INDEX>::gen_index(
-                i->origin[0],
-                i->origin[1],
-                i->origin[2]);
-            accessor.save(target, i->length(), offset, count);
+                static_cast<long>(i->origin[0]),
+                static_cast<long>(i->origin[1]),
+                static_cast<long>(i->origin[2]));
+            accessor.save(target, i->length(), offset, static_cast<std::size_t>(count));
 
             offset += i->length();
         }
